// src/common/enums/orderStatusStates.ts
export enum OrderStatusState {
	FORM_IN_PROGRESS = "Form-In-Progress",
	READY_FOR_VENDOR = "Ready-For-Vendor",
	READY_FOR_SYSTEM = "Ready-For-System",
	VENDOR_REVIEW_IN_PROGRESS = "Vendor-Review-In-Progress",
	ORDER_VALIDATION_COMPLETED = "Order-Validation-Completed",
	FIRM_ORDER_CONFIRMATION = "Firm-Order-Confirmation",
	CONNECTION_IN_PROGRESS = "Connection-In-Progress",
	COMPLETED = "Completed",
	REJECTED = "Rejected",
	CANCELLED = "Cancelled",
	CLOSED = "Closed",
}